Backend/NDN-live/src/requirements.txt
python-ndn==0.3
schedule==1.1
minio==7.1.2
requests==2.22.0
Flask==2.0.3
gunicorn==20.1.0
